pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labrat"
version = "0.2.0"
description = "Local-first runtime for long-running population search."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
nlp-sentiment = [
  "scikit-learn>=1.3",
  "scipy>=1.11",
  "datasets>=2.14",
  "catboost>=1.2",
  "lightgbm>=4.0",
]

[project.scripts]
labrat = "labrat.cli:main"

[tool.setuptools]
packages = ["labrat"]
